REFACTOR ALL THE THINGS.

+++ src/main/java/dk/itu/kelvin/math/MercatorProjection.java
@@ -10,7 +10,7 @@ public final class MercatorProjection implements Projection {
   /**
    * Factor with which to scale the lat/lon projection coordinates.
    */
-  private static final float PROJECTION_FACTOR = 30000;
+  private static final double PROJECTION_FACTOR = 30000;
 
   /**
    * Convert an x-coordinate into a spherical longitude.
@@ -21,7 +21,7 @@ public final class MercatorProjection implements Projection {
    * @param x The x-coordinate to convert.
    * @return  The corresponding spherical longitude.
    */
-  public float xToLon(final float x) {
+  public double xToLon(final double x) {
     return x / PROJECTION_FACTOR;
   }
 
@@ -34,7 +34,7 @@ public final class MercatorProjection implements Projection {
    * @param lon The spherical longitude to convert.
    * @return    The corresponding x-coordinate.
    */
-  public float lonToX(final float lon) {
+  public double lonToX(final double lon) {
     return lon * PROJECTION_FACTOR;
   }
 
@@ -47,8 +47,8 @@ public final class MercatorProjection implements Projection {
    * @param y The y-coordinate to convert.
    * @return  The corresponding spherical latitude.
    */
-  public float yToLat(final float y) {
-    return (float) Math.toDegrees(
+  public double yToLat(final double y) {
+    return (double) Math.toDegrees(
       2 * Math.atan(Math.exp(Math.toRadians(y))) - Math.PI / 2
     ) / -PROJECTION_FACTOR;
   }
@@ -62,8 +62,8 @@ public final class MercatorProjection implements Projection {
    * @param lat The spherical latitude to convert.
    * @return    The corresponding y-coordinate.
    */
-  public float latToY(final float lat) {
-    return (float) Math.toDegrees(
+  public double latToY(final double lat) {
+    return (double) Math.toDegrees(
       Math.log(Math.tan(Math.PI / 4 + Math.toRadians(lat) / 2))
     ) * -PROJECTION_FACTOR;
   }
